backend/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
pydantic==2.12.0
pydantic-settings==2.13.1
httpx==0.28.1
psycopg-binary==3.3.3
feedparser==6.0.12
python-dotenv==1.2.2
yfinance>=0.2.0
websockets>=12.0
